To create radio button feature which uses almost same logic with CheckBox, extract the common logic from CheckBox and make it as AbstractCheckButton class.
